requirements.txt

Django==3.1.7
django-celery==3.3.1
django-celery-beat==1.0.0
pyserial==2.5
redis==2.10.5
RPi.GPIO==0.5.11
